事件,手势,对应关系管理.

+++ Seals3Lab6/RLMGesture.h
@@ -9,9 +9,9 @@
 #import <Realm/Realm.h>
 
 @interface RLMGesture : RLMObject
-@property NSInteger objectId;
+@property int objectId;
 @property NSString *name;
-@property NSInteger type;
+@property int type;
 @end
 
 // This protocol enables typed collections. i.e.:
